backend/requirements.txt
Flask==2.2.3
gunicorn==20.1.1
Pillow==9.5.0
numpy==1.24.4
